Changed event rate from int to float.

+++ src/de/cco/jaer/eval/TrackerParams.java
@@ -122,7 +122,7 @@ public interface TrackerParams {
      * Getter for event rate.
      * @return Event rat in Hz.
      */
-    public int getEventRate();
+    public float getEventRate();
     
     /**
      * Print selected data
